Bug 57480: Fix some unnecessary casts, generics, Eclipse warnings, ...
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1658594 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFFont.java
@@ -309,7 +309,7 @@ public final class HSSFFont implements Font {
     {
         byte charset = font.getCharset();
         if(charset >= 0) {
-           return (int)charset;
+           return charset;
         } else {
            return charset + 256;
         }
